refactor actions & api-actions
