+++ project/src/store/action-type.ts
@@ -1,0 +1,24 @@
+/**
+ * @type {ActionType} - Enum Action for redux
+ * */
+export enum ActionType {
+  ChangeCity = 'city/change',
+  ChangeCommentLoadingStatus = 'comment/loadingStatus',
+  ChangeFavoriteLoadingStatus = 'favorite/loadingStatus',
+  ChangeFavoritePageLoadingStatus = 'offer/loadingPageStatus',
+  ChangeOfferFavoriteStatus = 'offer/changeFavoriteStatus',
+  ChangeOfferLoadingStatus = 'offer/loadingStatus',
+  ChangeSorting = 'offers/sorting',
+  ChangeUserInfo = 'user/info',
+  ClearOfferAction = 'offer/clearData',
+  ClearOffersFavoriteStatus = 'offer/clearOffersFavoriteStatus',
+  FillOffers = 'offers/fill',
+  LoadComments = 'data/comments',
+  LoadFavorites = 'data/favorites',
+  LoadNearbyOffers = 'data/nearbyOffers',
+  LoadOffer = 'data/offer',
+  LoadOffers = 'data/loadOffers',
+  RequireAuthorization = 'user/requireAuthorization',
+  RequireLogout = 'user/requireLogout',
+  RedirectToRoute = 'user/redirectToRoute'
+}
