+++ project/src/store/actions/index.ts
@@ -1,0 +1,31 @@
+export {
+  changeCommentLoadingStatusAction
+} from './comment';
+export {
+  changeFavoriteLoadingStatusAction,
+  changeFavoritePageLoadingStatusAction,
+  loadFavoritesAction
+} from './favorite';
+export {
+  changeOfferFavoriteStatusAction,
+  changeOfferLoadingStatusAction,
+  clearOfferAction,
+  loadCommentsAction,
+  loadNearbyOfferAction,
+  loadOfferAction
+} from './offer';
+export {
+  changeCityAction,
+  changeSortingAction,
+  clearOffersFavoriteStatusAction,
+  fillOffersAction,
+  loadOffersAction
+} from './offers';
+export {
+  redirectToRoute
+} from './route';
+export {
+  changeUserInfoAction,
+  requireAuthorization,
+  requireLogout
+} from './user';
