+++ project/src/store/api-actions/index.ts
@@ -1,0 +1,20 @@
+export {
+  addCommentAction,
+  fetchCommentsAction
+} from './comment';
+export {
+  changeFavoriteStatusOffer,
+  fetchFavoriteOffersAction
+} from './favorite';
+export {
+  fetchNearbyOfferAction,
+  fetchOfferAction
+} from './offer';
+export {
+  fetchOffersAction
+} from './offers';
+export {
+  checkAuthAction,
+  loginAction,
+  logoutAction
+} from './user';
